services/nlp-worker/requirements.txt
# Web Framework
fastapi
uvicorn
python-multipart

# AI & Data
google-genai
pandas
pydantic[email]

# Database
motor
dnspython

# Security & Auth
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==3.2.2

# Environment
python-dotenv